Resolved a buffer sync issue for SAMD platform.
The Wire.requestFrom() method on the AVR platform always
writes to the beginning of the RX buffer. The SAMD uses a ring
buffer and does not clear it before every request.

The change here reads the final byte of data sent from the
SHT2x, and so clears the buffer.